Moves AccountSecondaryIndexes types to submodule
